+++ packages/core/src/messages/requestSignature.ts
@@ -7,7 +7,7 @@ export type SignatureRequest = {
   message: string;
   address?: string;
   network?: string;
-  method?: 'aleo' | 'puzzle'
+  method?: 'aleo' | 'puzzle';
 };
 
 export type SignatureResponse = {
@@ -20,7 +20,7 @@ export const requestSignature = async ({
   message,
   address,
   network,
-  method
+  method,
 }: SignatureRequest): Promise<SignatureResponse> => {
   const connection = await getWalletConnectModalSignClient();
 
@@ -45,7 +45,7 @@ export const requestSignature = async ({
         params: {
           message,
           address: aleoAddressRegex.test(address ?? '') ? address : undefined,
-          method
+          method,
         } as SignatureRequest,
       },
     });