+++ packages/react/src/hooks/useAccount.ts
@@ -87,7 +87,7 @@ export const useAccount = () => {
             shortenedAddress: shortenAddress(data.address),
           });
         },
-        dependencies: []
+        dependencies: [],
       },
     ],
   });