+++ packages/react/src/hooks/useBalance.ts
@@ -69,7 +69,7 @@ export const useBalance = ({ address, multisig }: UseBalanceParams = {}) => {
           return !multisig;
         },
         onData: () => refetch(),
-        dependencies: [multisig]
+        dependencies: [multisig],
       },
       {
         subscriptionName: 'onSharedAccountSynced',
@@ -77,7 +77,7 @@ export const useBalance = ({ address, multisig }: UseBalanceParams = {}) => {
           return !!multisig && data?.address === address;
         },
         onData: () => refetch(),
-        dependencies: [multisig, address]
+        dependencies: [multisig, address],
       },
     ],
   });