+++ packages/react/src/hooks/useRecords.ts
@@ -92,7 +92,7 @@ export const useRecords = ({
         subscriptionName: 'onSelectedAccountSynced',
         condition: () => !multisig,
         onData: () => refetch(),
-        dependencies: [multisig]
+        dependencies: [multisig],
       },
       {
         subscriptionName: 'onSharedAccountSynced',
@@ -100,7 +100,7 @@ export const useRecords = ({
           return !!multisig && data?.address === address;
         },
         onData: () => refetch(),
-        dependencies: [multisig, address]
+        dependencies: [multisig, address],
       },
     ],
   });