+++ packages/react/src/hooks/utils/useInjectedSubscription.ts
@@ -10,7 +10,7 @@ type SubscriptionConfig = {
   subscriptionName: string;
   condition: (data: AccountSelectedResponse) => boolean;
   onData: (data: AccountSelectedResponse) => void;
-  dependencies: any[]
+  dependencies: any[];
 };
 
 type UseInjectedSubscriptionsParams = {
@@ -53,7 +53,7 @@ const useInjectedSubscriptions = ({
     // Cleanup on unmount or when dependencies change
     return () => {
       subscriptions.forEach((subscription) => {
-        subscription.unsubscribe()
+        subscription.unsubscribe();
       });
     };
   }, [session?.topic, ...configs.flatMap((config) => config.dependencies)]);