+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @puzzlehq/sdk
 
+## 0.5.3
+
+### Patch Changes
+
+- Adding tokenIds to createEventRequest
+- Updated dependencies
+  - @puzzlehq/sdk-core@0.5.3
+
 ## 0.5.2
 
 ### Patch Changes